cage: optionally return exitcode of primary client
Some applications indicate different shutdown conditions by returning
specific exit codes. One of these is e.g. Kodi, which returns 64 in case
the user chose "Power off" and 66 in case the user chose "Reboot".
In order to act on these exit codes, it thus makes sense in some
situations to pass them on from the primary client to the caller of
Cage.

Add a new flag "-e". If it's set and Cage shutso down because of an
orderly exit of its primary client, then it causes Cage to return the
primary client's exit code instead of its own one.