virtual keyboards: add virtual keyboards to seat list
Without adding the keyboard to seat list we wouldn't properly declare
keyboard capability, making the keyboard not work in headless mode.

We actually need to free this at some point so adding to the list is
most sensible, but we still do not want to group virtual keyboards with
regular ones so add a new 'is_virtual' bool to cg_keyboard_group so we
can skip these for grouping.